backend/src/user/requirements.txt
fastapi
uvicorn
supabase
pydantic
passlib[bcrypt]
python-consul
prometheus-fastapi-instrumentator
python-dotenv